weekly update master g98024d5a-1274 2024-12-02
98024d5a - [Fix](boot) Update cv180x fip.bin and boot0
38d3279d - [Fix](isp-tooldaemon) Fix capture yuv err
dae546bf - [Doc](manual) Update pqtool_manual.md
31df75bc - [Doc](manual) Update usb_cfg_manual.md
68c06007 - [Fix](protocol) Resolve packet sticking in sev3x
5c58ca8a - [Fix](pqtool) Decouple pqtool_uart with solution
1507447f - [Fix](middleware) Modification of CVI_U64/CVI_S64
0c703553 - [Fix](daemon2_uart) Optimize the timing of triggering console_init() when use uart for connecting pqtool
f0995b11 - [Feat](sns) Support sp2509 multiplex
71e29546 - [Feat](shell) Update isp compress script
0f5e7078 - [Fix](uvc) Fix vpss and venc frame rate decline
07b8f397 - [Fix](rgn) Modification of uintxx_t/intxx_t Data Types.
12fc1c9d - [Fix](fip.bin) Update fip.bin,build by v410
5f6e46be - [Feat](vpss) Change uintxx_t/intxx_t to CVI_UXX/CVI_SXX
4bd4e826 - [Fix](boot) Fix boot fail on 181xh board and update fip which is compiled by v4.1.0
d854bc23 - [Feat](OTA) Update cdc ota
fe79a8c2 - [Feat](base/sys) Change uintxx_t/intxx_t to CVI_UXX/CVI_SXX
8095dd25 - [Fix](vo) Change uintxx_t/intxx_t to CVI_UXX/CVI_SXX
d8852553 - [Feat](base) Modification of CVI_U64/CVI_S64
0edbcd92 - [Feat](boot) Updated fip.bin and boot for new spinor flash 2.fix cv181xh_boot build process
7c71c02b - [Feat](flash) Add ZB25VQ128D spinor flash
415ddc7e - [Fix](sns) Modify gc3003 bayer_format
0e98a8d4 - [Fix](venc) Fix unable to find fw_h26x error
6f26c532 - [Feat](sns) Add sensor gc3003
f01648eb - [Feat](tdl) Add CVI_TDL_Service_XXX support.
da1f99b6 - [Feat](sns) Add sc035hgs support
6efb3435 - [Feat](pqtool) Add connection ways of uart to pqtool

+++ components/cvi_mmf_sdk_cv180xx/include/cvi_middleware/include/cvi_type.h
@@ -43,10 +43,10 @@ typedef double                  CVI_DOUBLE;
 typedef void                    CVI_VOID;
 typedef bool                    CVI_BOOL;
 
-typedef uint64_t                CVI_U64;
-typedef int64_t                 CVI_S64;
+typedef unsigned long long int  CVI_U64;
+typedef long long int           CVI_S64;
 
-typedef size_t                  CVI_SIZE_T;
+typedef unsigned int            CVI_SIZE_T;
 
 /*----------------------------------------------
  * const defination
